moved all secrets based tests to smoke folder, updated npm scripts to properly run new smoke tests, and removed a few unused bin bash scripts
