+++ test/smoke/agent/start-stop.tap.js
@@ -6,15 +6,13 @@
 'use strict'
 
 const tap = require('tap')
-const configurator = require('../../lib/config')
-const Agent = require('../../lib/agent')
-const CollectorAPI = require('../../lib/collector/api')
-const {getTestSecret, shouldSkipTest} = require('../helpers/secrets')
+const configurator = require('../../../lib/config')
+const Agent = require('../../../lib/agent')
+const {getTestSecret} = require('../../helpers/secrets')
 
 
 const license = getTestSecret('TEST_LICENSE')
-const skip = shouldSkipTest(license)
-tap.test('no proxy set should not use proxy agent', {skip}, (t) => {
+tap.test('Collector API should connect to staging-collector.newrelic.com', (t) => {
   const config = configurator.initialize({
     app_name: 'node.js Tests',
     license_key: license,
@@ -24,7 +22,6 @@ tap.test('no proxy set should not use proxy agent', {skip}, (t) => {
     utilization: {
       detect_aws: false,
       detect_pcf: false,
-      detect_azure: false,
       detect_gcp: false,
       detect_docker: false
     },
@@ -33,18 +30,14 @@ tap.test('no proxy set should not use proxy agent', {skip}, (t) => {
     }
   })
   const agent = new Agent(config)
-  const api = new CollectorAPI(agent)
 
-
-  api.connect((error, response) => {
+  agent.start((error, returned) => {
     t.notOk(error, 'connected without error')
-
-    const returned = response && response.payload
     t.ok(returned, 'got boot configuration')
     t.ok(returned.agent_run_id, 'got run ID')
     t.ok(agent.config.run_id, 'run ID set in configuration')
 
-    api.shutdown((error) => {
+    agent.stop((error) => {
       t.notOk(error, 'should have shut down without issue')
       t.notOk(agent.config.run_id, 'run ID should have been cleared by shutdown')
