+++ test/smoke/api/connection.tap.js
@@ -8,11 +8,10 @@
 const tap = require('tap')
 const configurator = require('../../../lib/config')
 const Agent = require('../../../lib/agent')
-const {getTestSecret, shouldSkipTest} = require('../../helpers/secrets')
+const {getTestSecret} = require('../../helpers/secrets')
 
 const license = getTestSecret('TEST_LICENSE')
-const skip = shouldSkipTest(license)
-tap.test('Collector API should connect to staging-collector.newrelic.com', {skip}, (t) => {
+tap.test('Collector API should connect to staging-collector.newrelic.com', (t) => {
   const config = configurator.initialize({
     app_name: 'node.js Tests',
     license_key: license,
