+++ test/smoke/e2e/README.md
@@ -1,0 +1,2 @@
+This folder contains a very basic Express server that is instrumented with the Node.js agent.  It will run a test suite that just makes a http requests and verifies
+that it does not crash the server and properly responds.
