+++ test/smoke/e2e/express-server.js
@@ -5,7 +5,7 @@
 
 'use strict'
 
-require('../../index') // same as require('newrelic')
+require('../../../index') // same as require('newrelic')
 const express = require('express')
 
 const app = express()
