--- test/smoke/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "express-smoke",
-  "version": "0.0.0",
-  "description": "smoke test for express",
-  "main": "express.tap.js",
-  "scripts": {
-    "test": "node ./express.tap.js"
-  },
-  "author": "Wraithan <wraithan@newrelic.com>",
-  "license": "UNLICENSED",
-  "dependencies": {
-    "express": "^4.6.0"
-  },
-  "devDependencies": {
-    "chai": "^1"
-  },
-  "private": true
-}
