+++ test/smoke/proxy-api-connection-port.tap.js
@@ -13,7 +13,7 @@ const read = require('fs').readFileSync
 const configurator = require('../../lib/config')
 const Agent = require('../../lib/agent')
 const CollectorAPI = require('../../lib/collector/api')
-const {getTestSecret, shouldSkipTest} = require('../helpers/secrets')
+const {getTestSecret} = require('../helpers/secrets')
 const { SSL_HOST } = require('../lib/agent_helper')
 
 let port = 0
@@ -22,9 +22,8 @@ const SSL_CONFIG = {
   cert: read(join(__dirname, '../lib/self-signed-test-certificate.crt')),
 }
 const license = getTestSecret('TEST_LICENSE')
-const skip = shouldSkipTest(license)
 
-tap.test('setting proxy_port should use the proxy agent', {skip}, (t) => {
+tap.test('setting proxy_port should use the proxy agent', (t) => {
   const server = proxySetup(https.createServer(SSL_CONFIG))
 
   server.listen(0, () => {
